% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllOrganizations.R
\name{getAllOrganizations}
\alias{getAllOrganizations}
\title{getAllOrganizations}
\usage{
getAllOrganizations()
}
\value{
returns a data.frame of all organizations
}
\description{
This function is used to retrieve information on all 
registered organizations in your Zendesk organization
}
\details{
This function will return a data.frame containing all fields on every organization.  
The tags key will be returned within the data.frame as a list.
}
\examples{
\dontrun{

## This requires Zendesk authentication
organizations <- getAllOrganizations()
}
}
\references{
\url{http://developer.zendesk.com/documentation/rest_api/organizations.html}
}
\author{
Tanya Cashorali
}
