% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllSatisfactionRatings.R
\name{getAllSatisfactionRatings}
\alias{getAllSatisfactionRatings}
\title{getAllSatisfactionRatings}
\usage{
getAllSatisfactionRatings()
}
\value{
returns a data.frame of all satisfaction ratings ordered chronologically by created date, from oldest to newest.
}
\description{
This function is used to return all satisfaction ratings for your organization.
}
\details{
This function can only be used by Admins within your organization. Satisfaction 
Ratings are ordered chronologically by created date, from oldest to newest.
}
\examples{
\dontrun{

## This requires Zendesk authentication
satisfaction_ratings <- getAllSatisfactionRatings()
}
}
\references{
\url{http://developer.zendesk.com/documentation/rest_api/satisfaction_ratings.html}
}
\author{
John Hornbeck
}
