% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllTicketMetrics.R
\name{getAllTicketMetrics}
\alias{getAllTicketMetrics}
\title{getAllTicketMetrics}
\usage{
getAllTicketMetrics()
}
\value{
returns a data.frame of all ticket metrics for all tickets
}
\description{
This function is used to retrieve all ticket metrics from all tickets in your Zendesk organization
}
\details{
This function will return a data.frame containing all ticket metrics, 
such as latest_comment_added_at, first_resolution_time_in_minutes, etc. 
on every ticket.
}
\examples{
\dontrun{

## This requires Zendesk authentication
ticket_metrics <- getAllTicketMetrics()
}
}
\references{
\url{http://developer.zendesk.com/documentation/rest_api/ticket_metrics.html}
}
\author{
Tanya Cashorali
}
