% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllTickets.R
\name{getAllTickets}
\alias{getAllTickets}
\title{getAllTickets}
\usage{
getAllTickets()
}
\value{
returns a data.frame of all tickets ordered chronologically by created date, from oldest to newest.
}
\description{
This function is used to return all tickets stored in your organization.
}
\details{
This function can only be used by Admins within your organization. 
Tickets are ordered chronologically by created date, from oldest to newest.
}
\examples{
\dontrun{

## This requires Zendesk authentication
tickets <- getAllTickets()
}
}
\references{
\url{http://developer.zendesk.com/documentation/rest_api/tickets.html}
}
\author{
Tanya Cashorali
}
