% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllUsers.R
\name{getAllUsers}
\alias{getAllUsers}
\title{getAllUsers}
\usage{
getAllUsers()
}
\value{
returns a data.frame of all users
}
\description{
This function is used to retrieve information on all registered Zendesk users in your organization
}
\details{
This function will return a data.frame containing all fields on 
every user in your organization. The photo key will be returned 
within the data.frame as a list.
}
\examples{
\dontrun{

## This requires Zendesk authentication
zendesk('username', 'password', 'https://help.example.com')
users <- getAllUsers()
}
}
\references{
\url{http://developer.zendesk.com/documentation/rest_api/users.html}
}
\author{
Tanya Cashorali
}
