% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTicket.R
\name{getTicket}
\alias{getTicket}
\title{getTicket}
\usage{
getTicket(ticket.id)
}
\arguments{
\item{ticket.id}{A Zendesk ticket ID (e.g. 888)}
}
\value{
returns a data.frame of ticket information for the given ticket.id
}
\description{
This function is used to return ticket information for a given ticket ID
}
\details{
This function will return a data.frame of ticket information for a 
given ticket.id. The fields key will be returned within the data.frame 
as a list.
}
\examples{
\dontrun{

## This requires Zendesk authentication
ticket <- getTicket(ticket.id)
}
}
\references{
\url{http://developer.zendesk.com/documentation/rest_api/tickets.html}
}
\author{
Tanya Cashorali
}
