% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{unlistDataFrame}
\alias{unlistDataFrame}
\title{unlistDataFrame}
\usage{
unlistDataFrame(dataframe)
}
\arguments{
\item{dataframe}{A data.frame containing columns of lists to be unlisted.}
}
\value{
returns a data.frame
}
\description{
Utility function to unlist the columns of each data.frame where necessary.
}
\details{
This function will return a data.frame with the list type columns unlisted 
except where a column is a list of lists.
}
\examples{
\dontrun{

## This requires Zendesk authentication
unlistDataFrame(dataframe)
}
}
\references{
\url{http://developer.zendesk.com/documentation/rest_api/tickets.html}
}
\author{
Tanya Cashorali
}
