% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{zendesk_call}
\alias{zendesk_call}
\title{zendesk_call}
\usage{
zendesk_call(link, zd_call, ...)
}
\arguments{
\item{link}{the full api endpoint link to be called}

\item{zd_call}{the endpoint to call}

\item{...}{any valid parameter to be passed from zendesk_get}
}
\value{
a list of the resulting dataframe from one api call and its next link
}
\description{
A helper function for making api calls in zendesk_get funtion
}
\details{
This function will make one api call to the provided link and return a list
of the dataframe with the content of the api call as well as the next link
provided by the cull (will be null if there are no more pages)
}
