% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{zendesk_get}
\alias{zendesk_get}
\title{zendesk_get}
\usage{
zendesk_get(zd_call, ...)
}
\arguments{
\item{zd_call}{the api call (tickets, users, etc.)}

\item{...}{any valid parameters for the associated call}
}
\value{
returns a data.frame of the results from the given call
}
\description{
This is a a general purpose function for any valid Zendesk API endpoint
}
\details{
This function will return a data.frame of the results returned from the
zendesk API for the endpoint passed to it. Any valid parameters can be passed
to refine the request, for some endpoints there are required parameters.
}
\examples{
\dontrun{

## This requires Zendesk authentication
tickets <- zendesk_get('tickets')
}
}
