\name{zendeskR-package}
\alias{zendeskR-package}
\alias{zendeskR}
\docType{package}
\title{ R Wrapper for Zendesk API 
}
\description{
This package provides an R wrapper to the Zendesk API.
}
\details{
\tabular{ll}{
Package: \tab zendeskR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-07-11\cr
License: \tab Simplified BSD\cr
}
}
\author{
Tanya Cashorali
Maintainer: Tanya Cashorali <tanyacash@gmail.com>
}
\references{
http://developer.zendesk.com/documentation/rest_api/introduction.html
}
\keyword{ zendeskR }

