% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{zendeskR}
\alias{zendeskR}
\alias{zendeskR-package}
\title{R Wrapper for Zendesk API}
\description{
This package provides an R wrapper to the Zendesk API.
}
\details{
\tabular{ll}{
Package: \tab zendeskR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-07-11\cr
License: \tab Simplified BSD\cr
}
}
\references{
\url{http://developer.zendesk.com/documentation/rest_api/introduction.html}
}
\author{
Tanya Cashorali Tanya Cashorali: <tanyacash@gmail.com>
}
