% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zendesk.R
\name{zendesk}
\alias{zendesk}
\title{zendesk}
\usage{
zendesk(username, password, url)
}
\arguments{
\item{username}{Your Zendesk username.}

\item{password}{Your Zendesk password (or API token).}

\item{url}{Your organization's Zendesk URL (e.g. https://help.basho.com).}
}
\value{
Initializes a Zendesk.com API session. Will throw an error if all 3 parameters are not passed to the function.
}
\description{
This function is used to create a Zendesk.com API session.
}
\details{
This function will initialize a Zendesk.com API session.
}
\examples{
\dontrun{

## This requires Zendesk authentication
zendesk('username', 'password', 'https://help.basho.com')
}
}
\references{
\url{http://developer.zendesk.com/documentation/rest_api/introduction.html}
}
\author{
Tanya Cashorali
}
